ABXD Omega - the beginning
http://rosey.78.lt/

This is the 13th March commit and original release of the GitHub
repository, since MoonlightCapital started to screw with the repo and
make unwanted changes that I could not reverse back, so had to open a
new repository and delete the old one.

Registration passkey code is coming soon as a plugin, instead of being
incorporated into the main code. Other features can be discussed in the
official ABXD-Omega thread.

Hope you enjoy.